added 'modify/add/delete' features

+++ src/adminWindow/adminLibrarianWindow/AdminLibrarianController.java
@@ -45,6 +45,51 @@ public class AdminLibrarianController {
 
     @FXML
     public void handleModifyButton(ActionEvent actionEvent) {
+        Dialog<ButtonType> dialog = new Dialog<>();
+        dialog.initOwner(gridPane.getScene().getWindow());
+        dialog.setTitle("Modify Librarian Details");
+        FXMLLoader fxmlLoader = new FXMLLoader();
+        fxmlLoader.setLocation(getClass().getResource("/adminWindow/newUserDialogFxml.fxml"));
+
+        try {
+            dialog.getDialogPane().setContent(fxmlLoader.load());
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        dialog.getDialogPane().getButtonTypes().add(ButtonType.OK);
+        dialog.getDialogPane().getButtonTypes().add(ButtonType.CANCEL);
+
+        NewUserDialogController controller = fxmlLoader.getController();
+
+        int addressIndex = tableView.getSelectionModel().getSelectedIndex();
+        User oldLibrarian = null;
+        try {
+            oldLibrarian = librarians.get(addressIndex);
+        } catch (Exception e) {}
+
+        if (oldLibrarian != null) {
+            controller.setUser(oldLibrarian);
+            Optional<ButtonType> result = dialog.showAndWait();
+
+            if (result.isPresent() && result.get() == ButtonType.OK) {
+                User newLibrarian = controller.getUser(1);
+                if (newLibrarian != null) {
+                    if (UsersRepository.getInstance().updateUser(oldLibrarian, newLibrarian)) {
+                        oldLibrarian.setEmail(newLibrarian.getEmail());
+                        oldLibrarian.setPassword(newLibrarian.getPassword());
+                        oldLibrarian.setFirstName(newLibrarian.getFirstName());
+                        oldLibrarian.setLastName(newLibrarian.getLastName());
+                        tableView.refresh();
+                    } else {
+                        Alert alert = new Alert(Alert.AlertType.WARNING);
+                        alert.setTitle("Modify Librarian Details");
+                        alert.setHeaderText("The librarian details could not be modified!");
+                        alert.setContentText("Librarian with entered email already exists!");
+                        alert.showAndWait();
+                    }
+                }
+            }
+        }
     }
 
     @FXML
@@ -74,8 +119,16 @@ public class AdminLibrarianController {
 
             User librarian = controller.getUser(1);
 
-            UsersRepository.getInstance().addUser(librarian);
-            librarians.add(librarian);
+            if (UsersRepository.getInstance().addUser(librarian)) {
+                librarian.setUserId(UsersRepository.getInstance().getUserId(librarian));
+                librarians.add(librarian);
+            } else {
+                Alert alert = new Alert(Alert.AlertType.WARNING);
+                alert.setTitle("Add New Librarian");
+                alert.setHeaderText("The librarian was not added!");
+                alert.setContentText("Librarian with entered email already exists!");
+                alert.showAndWait();
+            }
         }
     }
 
@@ -84,8 +137,16 @@ public class AdminLibrarianController {
         try {
             int addressIndex = tableView.getSelectionModel().getSelectedIndex();
             User librarian = librarians.get(addressIndex);
-            UsersRepository.getInstance().deleteUser(librarian);
-            librarians.remove(addressIndex);
+
+            Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+            alert.setTitle("Delete Librarian");
+            alert.setHeaderText("Are you sure you want to delete the librarian: " + librarian.getFirstName() + " " + librarian.getLastName());
+            Optional<ButtonType> result  = alert.showAndWait();
+
+            if (result.isPresent() && result.get() == ButtonType.OK) {
+                UsersRepository.getInstance().deleteUser(librarian);
+                librarians.remove(addressIndex);
+            }
         } catch (Exception e) {
             System.out.println("Select an item");
         }
